Split CacheType into its own source file
