+++ lib/go-tc/enum/enum.go
@@ -52,20 +52,6 @@ type CacheGroupName string
 // DeliveryServiceName is the name of a CDN delivery service.
 type DeliveryServiceName string
 
-// CacheType is the type (or tier) of a CDN cache.
-type CacheType string
-
-const OriginLocationType = "ORG_LOC"
-
-const (
-	// CacheTypeEdge represents an edge cache.
-	CacheTypeEdge = CacheType("EDGE")
-	// CacheTypeMid represents a mid cache.
-	CacheTypeMid = CacheType("MID")
-	// CacheTypeInvalid represents an cache type enumeration. Note this is the default construction for a CacheType.
-	CacheTypeInvalid = CacheType("")
-)
-
 const AlgorithmConsistentHash = "consistent_hash"
 
 const MonitorTypeName = "RASCAL"
@@ -90,30 +76,6 @@ func (t TrafficMonitorName) String() string {
 
 func (d DeliveryServiceName) String() string {
 	return string(d)
-}
-
-// String returns a string representation of this cache type.
-func (t CacheType) String() string {
-	switch t {
-	case CacheTypeEdge:
-		return "EDGE"
-	case CacheTypeMid:
-		return "MID"
-	default:
-		return "INVALIDCACHETYPE"
-	}
-}
-
-// CacheTypeFromString returns a cache type object from its string representation, or CacheTypeInvalid if the string is not a valid type.
-func CacheTypeFromString(s string) CacheType {
-	s = strings.ToLower(s)
-	if strings.HasPrefix(s, "edge") {
-		return CacheTypeEdge
-	}
-	if strings.HasPrefix(s, "mid") {
-		return CacheTypeMid
-	}
-	return CacheTypeInvalid
 }
 
 // These are prefixed "QueryStringIgnore" even though the values don't always indicate ignoring, because the database column is named "qstring_ignore"
